+++ tests/test_sdba/test_detrending.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from xclim.sdba.detrending import PolyDetrend
+from xclim.sdba.detrending import LoessDetrend, PolyDetrend
 
 
 def test_poly_detrend(series):
@@ -15,4 +15,16 @@ def test_poly_detrend(series):
     # - The date is approximate (middle of the period)
     # - The last period may not be complete.
     np.testing.assert_array_almost_equal(dx, 0)
+    np.testing.assert_array_almost_equal(xt, x)
+
+
+def test_loess_detrend(series):
+    x = series(np.arange(20 * 365.25), "tas")
+    det = LoessDetrend(group="time", d=0, niter=1, f=0.2)
+    fx = det.fit(x)
+    dx = fx.detrend(x)
+    xt = fx.retrend(dx)
+
+    # Strong boundary effects in LOESS, remove ~ f * Nx on each side.
+    np.testing.assert_array_almost_equal(dx.isel(time=slice(1500, 5800)), 0)
     np.testing.assert_array_almost_equal(xt, x)